+++ Changelog.md
@@ -42,7 +42,10 @@ __New Features__
   - Miscellaneous improvements.
 - Adds more error-checking for inappropriate inputs (e.g., HVAC SHR=0 or clothes washer IMEF=0).
 - Allow alternative label energy use (W) input for ceiling fans.
-- Stochastic schedule for TV plug loads now use TV plug load schedules (instead of mirroring stochastic schedule for miscellaneous plug loads).
+- BuildResidentialScheduleFile measure:
+  - Allows appending columns to an existing CSV file rather than overwriting.
+  - TV plug load schedules now use TV schedule fractions and monthly multipliers (instead of mirroring miscellaneous plug load schedules).
+  - Ceiling fan monthly multipliers now use defaults based on monthly average outdoor temperatures per ANSI/RESNET/ICC 301-2019.
 
 __Bugfixes__
 - Fixes error if using AllowIncreasedFixedCapacities=true w/ HP detailed performance data.